add aggPrescription search
