+++ mobile.py
@@ -15,6 +15,7 @@ from routes.intervention import app_itrv
 from routes.static import app_stc
 from routes.substance import app_sub
 from routes.memory import app_mem
+from routes.patient import app_pat
 import logging
 import os
 
@@ -43,6 +44,7 @@ app.register_blueprint(app_itrv)
 app.register_blueprint(app_stc)
 app.register_blueprint(app_sub)
 app.register_blueprint(app_mem)
+app.register_blueprint(app_pat)
 
 CORS(app)
