+++ models/main.py
@@ -2,6 +2,7 @@ from flask_sqlalchemy import SQLAlchemy
 from sqlalchemy import func, text, and_, or_, desc, asc, distinct, cast
 from datetime import date, timedelta
 from sqlalchemy.dialects import postgresql
+from sqlalchemy.orm import deferred
 from routes.utils import *
 
 db = SQLAlchemy()
